eliminate conversion warning
git-svn-id: svn+ssh://cvs.lpds.sztaki.hu/var/lib/svn/szdg/dcapi/trunk@731 a7169a2c-3604-0410-bc95-c702d8d87f7a

+++ dcapi/common/cfg-client.c
@@ -156,7 +156,7 @@ int DC_getCfgInt(const char *key, int defaultValue)
 	/* Check for unit suffixes */
 	if (p && *p)
 	{
-		long mult = _DC_processSuffix(p);
+		long mult = (long) _DC_processSuffix(p);
 		if (mult == -1)
 		{
 			DC_log(LOG_WARNING, "Configuration value for key %s "
